+++ src/main/java/com/bibliomanager/library/repository/TypeRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
  @Repository: repository in the persistence layer and makes it eligible for Spring’s exception translation mechanism.
 
 */
-public interface TypeRepository extends CrudRepository<Type, Integer> {
+public interface TypeRepository extends CrudRepository<Type, Long> {
     @Query("SELECT COUNT(t) FROM Type t")
     long countTypes();
